requirements.txt
asn1crypto==0.22.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage==0.35.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pycparser==2.18
pydocumentdb==2.2.0
pygal==2.4.0
python-dateutil==2.6.1
requests==2.18.3
six==1.10.0
urllib3==1.22
Werkzeug==0.12.2